# openapi.yaml
{
  "openapi": "3.0.0",
  "info": {
    "title": "Spark ETL API",
    "version": "1.0.0",
    "description": "API documentation for the Spark ETL application, generated from server actions."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local API server"
    }
  ],
  "paths": {
...
    "/channel/{id}": {
      "get": {
        "summary": "Get Channel by ID",
        "description": "Retrieves a channel by its unique ID.",
        "operationId": "getChannelById",
        "tags": [
          "Channel"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Unique ID of the channel."
          }
        ],
        "responses": {
          "200": {
            "description": "Channel retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Channel",
        "description": "Updates an existing channel by its unique ID.",
        "operationId": "updateChannel",
        "tags": [
          "Channel"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Unique ID of the channel to update."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChannelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Channel updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Channel",
        "description": "Deletes a channel by its unique ID.",
        "operationId": "deleteChannel",
        "tags": [
          "Channel"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Unique ID of the channel to delete."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteChannelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Channel deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/community/{id}": {
      "get": {
        "summary": "Get community by ID",
        "description": "Retrieves a community by its unique ID.",
        "operationId": "getCommunityById",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Unique ID of the community."
          },
          {
            "in": "query",
            "name": "withSpaceUsers",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Include associated space users in the response."
          }
        ],
        "responses": {
          "200": {
            "description": "Community retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Community"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a community",
        "description": "Updates an existing community record by its ID.",
        "operationId": "updateCommunity",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The ID of the community to update."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommunityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Community updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Community"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a community",
        "description": "Deletes a community record by its ID.",
        "operationId": "deleteCommunity",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The ID of the community to delete."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCommunityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Community deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}